hf_publish/space/requirements.txt
timm==1.0.26
torch
torchvision
gradio>=4.0.0
huggingface_hub>=0.32.0
pillow
